% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{detectEncoding}
\alias{detectEncoding}
\title{Detect encoding of input string}
\usage{
detectEncoding(str)
}
\arguments{
\item{str}{Character vector.}
}
\value{
Character vector with encoding names.
}
\description{
Detect encoding of input string
}
\references{
\url{https://www.freedesktop.org/wiki/Software/uchardet/}
}
